myshop/requirements.txt
# pip install -r requirements.txt
Django==4.2.7
django-debug-toolbar==4.2.0
python-dotenv==1.0.0
psycopg2-binary==2.9.9
celery==5.3.6
redis==5.0.1
pylint==3.0.3
pandas==2.1.4
openpyxl==3.1.2
pydantic==2.6.0
djangorestframework==3.14.0